Implement a simple stufftext filter.
Apparently there are malicious servers in the wild which exploits
stufftexts in attacks to several Quake II clients. So far I haven't
a report of the whole process getting hijacked, just cvars getting
overwritten. However given the state of Quake IIs command parser
anything is possible...

When `cl_stuftext 0` only known to be good stufftexts which are
required for baseq2, rogue and xatrix are allowed. Everything else
is silently discarded. I cannot be 100% sure that this prevents
all possible attacks, this may even a variant of the halting
problem... However it should be good enough to reduce the attack
surface significantly.

Since this will break any server or mod which required random
stufftexts `cl_stufftext 1` will stay the default.

Many thanks to Asa Cherub for the report.