Set the `ambush` spawnflag on monster_boss2 in biggun.
I encountered a bug where the monster_boss2 aka hornet in biggun.bsp
somehow slipped into the void while pushed out of it's compartment.
I'm unable to reproduce it, but analysis showed that the hornet isn't
spawned in ambush mode. As soon as the player fires a gun it starts
moving. Maybe that is enough to slip out of the compartment.

Additionally the hornet is a targeted monster. Monsters triggered by
level logic must be set to ambush, otherwise sound and level logic
triggering mess each other up.

Closes #1078.