sdl3: Fix SDL_GetFullscreenDisplayModes const-ness
Previously this returned a const pointer which we weren't supposed to
free (unlike in SDL2). This was part of the now-removed temporary
storage thing that SDL3 had for awhile to stop the user from having to
free memory.

This did mean there was previously a double-free bug here, since we've
always been freeing this ourselves.

This tracks SDL3 commit 4f55271571fb.