Fix drawing of UDP download progress bar with `r_consolescale` > `1`.
When calculating the Y position of the UDP download progress bar the
scale factor must be taken into account. Otherwise the bar will be drawn
outside the visible area of the console. This was missed when scaling
was first implemented years ago and broken since then.

Reported by @SiemensSchuckert, closes #1172.