"Fix" window positioning with SDL 3.
Window position handling works slightly different on SDL 3. On SDL 2 we
could assume, that window identifiers started at 0 and increased by 1
with every display. While not documented that way it worked. With SDL 3
window identifiers are random, depending on the underlying window
system. Work around that by keeping the window identifiers in an array
and use the array indices as user exposed identifiers. With this the
user still sees 0 for the first display, 1 for the second, etc.

In my opinion window position handling und fundamentally flawed and
would need a major refactoring of the GLimp backends to work properly.
Currently the window position can be changed and any time by setting
`vid_displayindex`. That is even exposed into the menu. However the
windows position is only applied when the window is created. A simple
`vid_restart` doesn't recreate the window, so the game must be
effectively restarted... Additionally not all window system allow
to set the window position. One example is Wayland.

Last but not least the window position handling doesn't integrate with
resolution setting and fullscreen handling. In might work in some
scenarios but it's definitely not safe.

Closes #1202.