Remove 'window index' menu setting.
The menu setting only appeared when SDL detected more than one display,
which is ... problematic. It set the `vid_displayindex` cvar, which is
fundamentally flawed and likely doesn't behave how the user expects. See
the last commit for details. That we never got a bug report for the cvar
or the menu option is a string hint that no one uses it. So remove the
menu setting, but keep the cvar.

As a nice side effect this makes one more menu slot available.

If someone really uses this functionality, a pull request fixing
`vid_displayindex` at least for SDL 3 is welcome.